+++ backend/src/workers/later-worker.ts
@@ -7,7 +7,7 @@ import GeorgeQueue from "../queues/george-queue";
 import { buildMonkeyMail } from "../utils/monkey-mail";
 import { DailyLeaderboard } from "../utils/daily-leaderboards";
 import { getCachedConfiguration } from "../init/configuration";
-import { formatSeconds, getOrdinalNumberString, mapRange } from "../utils/misc";
+import { formatSeconds, getOrdinalNumberString } from "../utils/misc";
 import LaterQueue, {
   type LaterTask,
   type LaterTaskContexts,
@@ -16,6 +16,7 @@ import LaterQueue, {
 import { recordTimeToCompleteJob } from "../utils/prometheus";
 import { WeeklyXpLeaderboard } from "../services/weekly-xp-leaderboard";
 import { MonkeyMail } from "@monkeytype/contracts/schemas/users";
+import { mapRange } from "@monkeytype/util/numbers";
 
 async function handleDailyLeaderboardResults(
   ctx: LaterTaskContexts["daily-leaderboard-results"]
