+++ frontend/src/ts/config.ts
@@ -24,11 +24,11 @@ import {
 } from "./utils/misc";
 import * as ConfigSchemas from "@monkeytype/contracts/schemas/configs";
 import { Config } from "@monkeytype/contracts/schemas/configs";
-import { roundTo1 } from "./utils/numbers";
 import { Mode, ModeSchema } from "@monkeytype/contracts/schemas/shared";
 import { Language, LanguageSchema } from "@monkeytype/contracts/schemas/util";
 import { LocalStorageWithSchema } from "./utils/local-storage-with-schema";
 import { migrateConfig } from "./utils/config";
+import { roundTo1 } from "@monkeytype/util/numbers";
 
 const configLS = new LocalStorageWithSchema({
   key: "config",
