+++ frontend/src/ts/controllers/ad-controller.ts
@@ -1,5 +1,5 @@
 import { debounce } from "throttle-debounce";
-// import * as Numbers from "../utils/numbers";
+// import * as Numbers from "@monkeytype/util/numbers";
 import * as ConfigEvent from "../observables/config-event";
 import * as BannerEvent from "../observables/banner-event";
 import Config from "../config";
