+++ frontend/src/ts/elements/profile.ts
@@ -2,7 +2,7 @@ import * as DB from "../db";
 import { format } from "date-fns/format";
 import { differenceInDays } from "date-fns/differenceInDays";
 import * as Misc from "../utils/misc";
-import * as Numbers from "../utils/numbers";
+import * as Numbers from "@monkeytype/util/numbers";
 import * as Levels from "../utils/levels";
 import * as DateTime from "../utils/date-and-time";
 import { getHTMLById } from "../controllers/badge-controller";
@@ -12,6 +12,7 @@ import { formatDistanceToNowStrict } from "date-fns/formatDistanceToNowStrict";
 import { getHtmlByUserFlags } from "../controllers/user-flag-controller";
 import Format from "../utils/format";
 import { UserProfile, RankAndCount } from "@monkeytype/contracts/schemas/users";
+import { abbreviateNumber, convertRemToPixels } from "../utils/numbers";
 
 type ProfileViewPaths = "profile" | "account";
 type UserProfileOrSnapshot = UserProfile | MonkeyTypes.Snapshot;
@@ -416,7 +417,7 @@ export function updateNameFontSize(where: ProfileViewPaths): void {
   const nameFieldjQ = details.find(".user");
   const nameFieldParent = nameFieldjQ.parent()[0];
   const nameField = nameFieldjQ[0];
-  const upperLimit = Numbers.convertRemToPixels(2);
+  const upperLimit = convertRemToPixels(2);
 
   if (!nameField || !nameFieldParent) return;
 
@@ -450,6 +451,6 @@ function formatXp(xp: number): string {
   if (xp < 1000) {
     return Math.round(xp).toString();
   } else {
-    return Numbers.abbreviateNumber(xp);
+    return abbreviateNumber(xp);
   }
 }
