+++ frontend/src/ts/elements/result-batches.ts
@@ -1,8 +1,8 @@
 import * as DB from "../db";
 import * as ServerConfiguration from "../ape/server-configuration";
-import { mapRange } from "../utils/misc";
 import { blendTwoHexColors } from "../utils/colors";
 import * as ThemeColors from "../elements/theme-colors";
+import { mapRange } from "@monkeytype/util/numbers";
 
 export function hide(): void {
   $(".pageAccount .resultBatches").addClass("hidden");
