+++ frontend/src/ts/test/caret.ts
@@ -1,10 +1,10 @@
-import * as Numbers from "../utils/numbers";
 import * as JSONData from "../utils/json-data";
 import Config from "../config";
 import * as TestInput from "./test-input";
 import * as SlowTimer from "../states/slow-timer";
 import * as TestState from "../test/test-state";
 import * as TestWords from "./test-words";
+import { convertRemToPixels } from "../utils/numbers";
 
 export let caretAnimating = true;
 const caret = document.querySelector("#caret") as HTMLElement;
@@ -164,7 +164,7 @@ export async function updatePosition(noAnim = false): Promise<void> {
   const letterHeight =
     currentLetter?.offsetHeight ||
     lastWordLetter?.offsetHeight ||
-    Config.fontSize * Numbers.convertRemToPixels(1);
+    Config.fontSize * convertRemToPixels(1);
 
   const letterPosTop =
     currentLetter?.offsetTop ?? lastWordLetter?.offsetTop ?? 0;
