+++ frontend/src/ts/test/test-ui.ts
@@ -11,7 +11,6 @@ import * as Replay from "./replay";
 import * as Misc from "../utils/misc";
 import * as Strings from "../utils/strings";
 import * as JSONData from "../utils/json-data";
-import * as Numbers from "../utils/numbers";
 import { get as getTypingSpeedUnit } from "../utils/typing-speed-units";
 import * as SlowTimer from "../states/slow-timer";
 import * as CompositionState from "../states/composition";
@@ -31,6 +30,7 @@ import {
   TimerColor,
   TimerOpacity,
 } from "@monkeytype/contracts/schemas/configs";
+import { convertRemToPixels } from "../utils/numbers";
 
 async function gethtml2canvas(): Promise<typeof import("html2canvas").default> {
   return (await import("html2canvas")).default;
@@ -456,7 +456,7 @@ export async function updateWordsInputPosition(initial = false): Promise<void> {
   const activeWordMargin =
     parseInt(computed.marginTop) + parseInt(computed.marginBottom);
 
-  const letterHeight = Numbers.convertRemToPixels(Config.fontSize);
+  const letterHeight = convertRemToPixels(Config.fontSize);
   const targetTop =
     activeWord.offsetTop + letterHeight / 2 - el.offsetHeight / 2 + 1; //+1 for half of border
 
@@ -517,7 +517,7 @@ function updateWordsHeight(force = false): void {
     }
     $(".outOfFocusWarning").css(
       "margin-top",
-      wordHeight + Numbers.convertRemToPixels(1) / 2 + "px"
+      wordHeight + convertRemToPixels(1) / 2 + "px"
     );
   } else {
     let finalWordsHeight: number, finalWrapperHeight: number;
@@ -577,7 +577,7 @@ function updateWordsHeight(force = false): void {
       $("#wordsWrapper").css("height", finalWrapperHeight + "px");
       $(".outOfFocusWarning").css(
         "margin-top",
-        finalWrapperHeight / 2 - Numbers.convertRemToPixels(1) / 2 + "px"
+        finalWrapperHeight / 2 - convertRemToPixels(1) / 2 + "px"
       );
     }, 0);
   }
@@ -698,8 +698,8 @@ export async function screenshot(): Promise<void> {
     true
   ) as number; /*clientHeight/offsetHeight from div#target*/
   try {
-    const paddingX = Numbers.convertRemToPixels(2);
-    const paddingY = Numbers.convertRemToPixels(2);
+    const paddingX = convertRemToPixels(2);
+    const paddingY = convertRemToPixels(2);
 
     const canvas = await (
       await gethtml2canvas()
