+++ frontend/src/ts/utils/ip-addresses.ts
@@ -1,4 +1,4 @@
-import { randomIntFromRange } from "../utils/numbers";
+import { randomIntFromRange } from "@monkeytype/util/numbers";
 
 function getRandomIPvXaddress(
   bits: number,
